mcpsynology/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synology-mcp"
version = "0.1.0"
description = "MCP server for Synology DSM NAS management"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Nicolas" }]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "keyring>=25.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.21.0",
    "mypy>=1.9.0",
    "ruff>=0.4.0",
]

[project.scripts]
synology-mcp = "synology_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/synology_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
